+++ dev/user.clj
@@ -2,7 +2,6 @@
   (:require [clojure.tools.namespace.repl :refer [refresh]]
             [clojure.walk :refer [postwalk]]
             [parcera.core :as parcera]
-            [instaparse.core :as insta]
             [grape.core :as g]
             [grape.cli :as cli]))